Fixed config names
